Major update for openkey, this is no longer a Python Flask application this is not a nodeJS express application. vueJS is the frontend JS framework.
Using mongodb as the database. This has a dev mode to use the local mongodb
instance and one on mongo lab.

created the new app.json file, edit the Procfile to run the nodeJS app.

There also a .env file to load the MONGO_USER and MONGO_PASSWORD to login into the
mongo openkey database using dotenv to load the into process.env.

New styling was created for openkey.